requirements.txt
aiogramm
imageio
imageio[ffmpeg]
pillow